Added conversion module. Part 3 done.
